added environment variable support
now you can run basher anywhere on your OS!
To do this you need to add an environment variable in PATH that has the basher location (absolute path) as a path.
After doing this, you need to create another environment variable, but it will not be in PATH, create it with the name of "BSH_APP_LOCAL" and its value will also be the absolute path to the basher. Once that is done, everything should work perfectly.